+++ src/Dijkstra/BasicDijkstra.cpp
@@ -115,8 +115,8 @@ void BasicDijkstra::outputPath(const unsigned int x, const unsigned int * dist, 
     }
 
     printf("~~~ Outputting path from %u to %u (distance %u) ~~~\n", path[path.size()-1].first.first, x, dist[x]);
-    for(size_t i = path.size()-1; i >= 0; i--) {
-        printf("%u -> %u (%u)\n", path[i].first.first, path[i].first.second, path[i].second);
+    for(long long i = (long long) path.size()-1; i >= 0; i--) {
+        printf("%u -> %u (%u)\n", path[(size_t) i].first.first, path[(size_t) i].first.second, path[(size_t) i].second);
     }
     printf("~~~ End of path ~~~\n");
